refactor: link dex mons to base data
