+++ src/components/panels/ActiveShlagemon.vue
@@ -9,16 +9,16 @@ const dex = useShlagedexStore()
   <div v-if="dex.activeShlagemon" class="flex items-center gap-2">
     <div class="h-16 w-16 flex-shrink-0">
       <img
-        :src="`/shlagemons/${dex.activeShlagemon.id}/${dex.activeShlagemon.id}.png`"
-        :alt="dex.activeShlagemon.name"
+        :src="`/shlagemons/${dex.activeShlagemon.base.id}/${dex.activeShlagemon.base.id}.png`"
+        :alt="dex.activeShlagemon.base.name"
         class="h-full w-full object-contain"
       >
     </div>
 
     <div class="info flex flex-col">
-      <span class="font-bold">{{ dex.activeShlagemon.name }}</span>
+      <span class="font-bold">{{ dex.activeShlagemon.base.name }}</span>
       <span>Lvl {{ dex.activeShlagemon.lvl }}</span>
-      <ShlagemonType :value="dex.activeShlagemon.type" />
+      <ShlagemonType :value="dex.activeShlagemon.base.type" />
     </div>
   </div>
 </template>
