+++ src/components/panels/SelectedShlagemon.vue
@@ -6,6 +6,6 @@ const dex = useShlagedexStore()
 
 <template>
   <div v-if="dex.activeShlagemon" class="h-full flex items-center justify-center">
-    <img :src="`/shlagemons/${dex.activeShlagemon.id}/${dex.activeShlagemon.id}.png`" :alt="dex.activeShlagemon.name" class="max-h-40 object-contain">
+    <img :src="`/shlagemons/${dex.activeShlagemon.base.id}/${dex.activeShlagemon.base.id}.png`" :alt="dex.activeShlagemon.base.name" class="max-h-40 object-contain">
   </div>
 </template>
